Move tests to Controller test classes
