+++ DevelopGitHub/GitHubExtension/GitHubExtension.Activity.Internal.Tests/GitHubExtension.Activity.Internal.Tests.csproj
@@ -91,12 +91,10 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Extensions\IdentityExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TestsForControllers\CommandControllerAddJoinToSystemActivityTests.cs" />
-    <Compile Include="TestsForControllers\CommandControllerRepositoryAddedToSystem.cs" />
-    <Compile Include="TestsForControllers\QueryControllerGetUserActivitiesRepositoryTests.cs" />
-    <Compile Include="TestsForControllers\QueryControllerGetUserActivitiesTests.cs" />
-    <Compile Include="TestsForControllers\CommandControllerAddRoleActivityTests.cs" />
+    <Compile Include="TestsForControllers\InternalActivityCommandControllerTests.cs" />
+    <Compile Include="TestsForControllers\InternalActivityQueryControllerTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\GitHubExtension.Activity.DAL\GitHubExtension.Activity.DAL.csproj">
